legaledge_backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
python-decouple>=3.8
cryptography>=42.0.0
psycopg[binary]>=3.2
